+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # Change Log
 
-## Master
+## 1.5 [2017-03-30]
 
 ## Breaking